Slight refactor in how env properties are passed down to the various classes
